Create Laboratory_work.7
This programm works in dosbox with TASM compiler. It uses an overlay Arifm.exe to do 
math calculations (+ - * /). First of all you must input a quantity of numbers you want to read from a file (file must contain a string like 2+9/8-7*2-8..... and so on ) using a command line (for example: if you input 3 you will get a string: 2+9/8). After that it generates a reverse polish notation and calculates it. 
At the end it displays an expression that was read from a file, a reverse polish notation of this expression and a result. Also it can display an error if something is incorrect in the file or in the command line.