Create Laboratory_work.8
This is an alarm clock resident program, that works in FreeDOS, TASM compiler. It uses a command line for input a time for an alarm clock and sound length in milliseconds. Then it checkes a string from the command line and transforms it into numbers. After that it changes 1ah interrupt on procedure that creates sound and writes "Alarm clock!". This programm is resident and after calling lefts in system untill alarm clock triggers.